refactor dates
let FE handle date format change(previously done on BE) from dd/mm/yyyy to yyyy-mm-dd for FILE UPLOAD feature

rename dateStringGenerator.js to date.js and move all date related functions into date.js

let FE handle dateformat presentation for upload success summary (previously done on BE)
